Resolves getExistingTag circular dependency.
